+++ tests/testthat/test-GDALRaster-class.R
@@ -6,6 +6,17 @@ test_that("info() prints output to the console", {
 	ds$close()
 })
 
+test_that("infoAsJSON() returns string output", {
+	evt_file <- system.file("extdata/storml_evt.tif", package="gdalraster")
+	tmp_file <- paste0(tempdir(), "/", "storml_evt_tmp.tif")
+	file.copy(evt_file,  tmp_file)
+	ds <- new(GDALRaster, tmp_file, TRUE)
+	expect_type(ds$infoAsJSON(), "character")
+	files <- ds$getFileList()
+	on.exit(unlink(files))
+	ds$close()
+})
+
 test_that("dataset parameters are correct", {
 	lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
 	ds <- new(GDALRaster, lcp_file, TRUE)
@@ -62,7 +73,6 @@ test_that("open/close/re-open works", {
 	r <- read_ds(ds)
 	ds$close()
 	mod_file <- paste0(tempdir(), "/", "storml_elev_mod.tif")
-	on.exit(unlink(mod_file))
 	rasterFromRaster(srcfile = elev_file,
 					dstfile = mod_file,
 					nbands = 1,
@@ -82,6 +92,8 @@ test_that("open/close/re-open works", {
 	ds$write(band=1, 0, 0, dm[1], dm[2], r)
 	expect_false(all(is.na(read_ds(ds))))
 	expect_true(any(is.na(read_ds(ds))))
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
@@ -89,16 +101,16 @@ test_that("statistics are correct", {
 	elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
 	mod_file <- paste0(tempdir(), "/", "storml_elev_mod.tif")
 	file.copy(elev_file,  mod_file)
-	on.exit(unlink(mod_file))
 	ds <- new(GDALRaster, mod_file, read_only=TRUE)
 	stats <- round(ds$getStatistics(band=1, approx_ok=FALSE, force=TRUE))
 	expect_equal(stats, c(2438, 3046, 2676, 133))
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
 test_that("floating point I/O works", {
 	f <- paste0(tempdir(), "/", "testfloat.tif")
-	on.exit(unlink(f))
 	create(format="GTiff", dst_filename=f, xsize=10, ysize=10,
 			nbands=1, dataType="Float32")
 	ds <- new(GDALRaster, f, read_only=FALSE)
@@ -116,12 +128,13 @@ test_that("floating point I/O works", {
 	ds$deleteNoDataValue(band=1)
 	ds$open(read_only=TRUE)
 	expect_equal(read_ds(ds), rep(-99999, 100))
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
 test_that("complex I/O works", {
 	f <- paste0(tempdir(), "/", "testcomplex.tif")
-	on.exit(unlink(f))
 	create(format="GTiff", dst_filename=f, xsize=10, ysize=10,
 			nbands=1, dataType="CFloat32")
 	ds <- new(GDALRaster, f, read_only=FALSE)
@@ -130,6 +143,8 @@ test_that("complex I/O works", {
 	ds$write(band=1, xoff=0, yoff=0, xsize=10, ysize=10, z)
 	ds$open(read_only=TRUE)
 	expect_vector(read_ds(ds), ptype=complex(0), size=100)
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
@@ -137,7 +152,6 @@ test_that("set unit type, scale and offset works", {
 	elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
 	mod_file <- paste0(tempdir(), "/", "storml_elev_mod.tif")
 	file.copy(elev_file,  mod_file)
-	on.exit(unlink(mod_file))
 	ds <- new(GDALRaster, mod_file, read_only=FALSE)
 	ds$setUnitType(1, "m")
 	ds$setScale(1, 1)
@@ -145,6 +159,8 @@ test_that("set unit type, scale and offset works", {
 	expect_equal(ds$getUnitType(1), "m")
 	expect_equal(ds$getScale(1), 1)
 	expect_equal(ds$getOffset(1), 0)
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
@@ -152,17 +168,17 @@ test_that("build overviews runs without error", {
 	elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
 	mod_file <- paste0(tempdir(), "/", "storml_elev.tif")
 	file.copy(elev_file,  mod_file)
-	on.exit(unlink(mod_file))
 	ds <- new(GDALRaster, mod_file, read_only=FALSE)
 	expect_no_error(ds$buildOverviews("BILINEAR", c(2,4,8), 0))
 	expect_no_error(ds$buildOverviews("NONE", 0, 0))
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
 test_that("get/set color table works", {
 	f <- system.file("extdata/storml_evc.tif", package="gdalraster")
 	f2 <- paste0(tempdir(), "/", "storml_evc_ct.tif")
-	on.exit(unlink(f2))
 	calc("A", f, dstfile=f2, dtName="UInt16", nodata_value=32767,
 			setRasterNodataValue=TRUE)
 	ds <- new(GDALRaster, f2, read_only=FALSE)
@@ -173,12 +189,13 @@ test_that("get/set color table works", {
 	evc_ct <- ds$getColorTable(1)
 	expect_equal(nrow(evc_ct), 400)
 	expect_equal(ds$getPaletteInterp(1), "RGB")
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })
 
 test_that("get/set band color interpretation works", {
 	f <- paste0(tempdir(), "/", "test_col_interp.tif")
-	on.exit(unlink(f))
 	create(format="GTiff", dst_filename=f, xsize=10, ysize=10,
 			nbands=1, dataType="Byte")
 	ds <- new(GDALRaster, f, read_only=FALSE)
@@ -215,5 +232,7 @@ test_that("get/set band color interpretation works", {
 	expect_equal(ds$getRasterColorInterp(1), "YCbCr_Cr")
 	ds$setRasterColorInterp(1, "Gray")
 	expect_equal(ds$getRasterColorInterp(1), "Gray")
+	files <- ds$getFileList()
+	on.exit(unlink(files))
 	ds$close()
 })